TUP-5205: do some refactor to move the display value calculate code to the util class to reuse by the compare editor
