+++ main/plugins/org.talend.designer.core/src/main/java/org/talend/designer/core/ui/editor/properties/controllers/AbstractRepositoryController.java
@@ -16,8 +16,6 @@ import java.beans.PropertyChangeEvent;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.eclipse.core.runtime.Assert;
-import org.eclipse.emf.common.util.EList;
 import org.eclipse.gef.commands.Command;
 import org.eclipse.jface.fieldassist.DecoratedField;
 import org.eclipse.jface.fieldassist.FieldDecoration;
@@ -36,36 +34,22 @@ import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Text;
 import org.eclipse.ui.views.properties.tabbed.ITabbedPropertyConstants;
-import org.talend.commons.exception.PersistenceException;
-import org.talend.commons.ui.runtime.exception.ExceptionHandler;
 import org.talend.commons.ui.runtime.image.ImageProvider;
-import org.talend.core.GlobalServiceRegister;
-import org.talend.core.IESBService;
 import org.talend.core.model.metadata.builder.connection.Connection;
-import org.talend.core.model.metadata.builder.connection.DatabaseConnection;
-import org.talend.core.model.metadata.builder.connection.QueriesConnection;
-import org.talend.core.model.metadata.builder.connection.Query;
 import org.talend.core.model.metadata.builder.database.ExtractMetaDataUtils;
 import org.talend.core.model.process.EParameterFieldType;
 import org.talend.core.model.process.IElementParameter;
 import org.talend.core.model.properties.ConnectionItem;
-import org.talend.core.model.properties.DatabaseConnectionItem;
-import org.talend.core.model.properties.FileItem;
 import org.talend.core.model.properties.Item;
-import org.talend.core.model.properties.LinkRulesItem;
-import org.talend.core.model.properties.RulesItem;
-import org.talend.core.model.repository.ERepositoryObjectType;
 import org.talend.core.model.repository.IRepositoryViewObject;
 import org.talend.core.ui.CoreUIPlugin;
 import org.talend.core.ui.properties.tab.IDynamicProperty;
-import org.talend.designer.core.DesignerPlugin;
 import org.talend.designer.core.model.components.EParameterName;
 import org.talend.designer.core.model.components.EmfComponent;
 import org.talend.designer.core.ui.editor.cmd.ChangeMetadataCommand;
 import org.talend.designer.core.ui.editor.nodes.Node;
 import org.talend.designer.core.ui.editor.properties.controllers.creator.SelectAllTextControlCreator;
 import org.talend.repository.UpdateRepositoryUtils;
-import org.talend.repository.model.IProxyRepositoryFactory;
 
 /**
  * DOC nrousseau class global comment. Detailled comment
@@ -76,6 +60,8 @@ public abstract class AbstractRepositoryController extends AbstractElementProper
 
     protected static final int STANDARD_REPOSITORY_WIDTH = 250;
 
+    protected ControllerRepositoryValueHander repositoryValueHander;
+
     /**
      * DOC nrousseau AbstractRepositoryController constructor comment.
      *
@@ -83,6 +69,7 @@ public abstract class AbstractRepositoryController extends AbstractElementProper
      */
     public AbstractRepositoryController(IDynamicProperty dp) {
         super(dp);
+        repositoryValueHander = new ControllerRepositoryValueHander();
     }
 
     protected abstract String getRepositoryTypeParamName();
@@ -321,204 +308,8 @@ public abstract class AbstractRepositoryController extends AbstractElementProper
         return initialSize.y + ITabbedPropertyConstants.VSPACE;
     }
 
-    protected ConnectionItem lastItemUsed;
-
-    private FileItem lastFileItemUsed; // hywang add for 6484
-
-    private LinkRulesItem lastLinkItem;
-
-    private void fastRepositoryUpdateProperty(IElementParameter param) {
-        if (param != null && param.getValue() != null) {
-            IProxyRepositoryFactory factory = DesignerPlugin.getDefault().getProxyRepositoryFactory();
-            String linkedRepository = (String) param.getValue();
-            // value stored is the id, so we can get this item directly
-            Item item;
-            String displayName = "";
-            try {
-                IRepositoryViewObject object = factory.getLastVersion(linkedRepository.split(" - ")[0]);
-                if (object == null) {
-                    return;
-                }
-                item = object.getProperty().getItem();
-                // Assert.isTrue(item instanceof ConnectionItem);
-                IESBService service = null;
-                if (GlobalServiceRegister.getDefault().isServiceRegistered(IESBService.class)) {
-                    service = (IESBService) GlobalServiceRegister.getDefault().getService(IESBService.class);
-                }
-                if (service != null && ERepositoryObjectType.getItemType(item) == service.getServicesType()) {
-                    lastItemUsed = (ConnectionItem) item;
-                    displayName = "Service:" + service.getServiceLabel(item, linkedRepository);
-                } else if (item instanceof ConnectionItem) {
-                    lastItemUsed = (ConnectionItem) item;
-                    displayName = dynamicProperty.getRepositoryAliasName(lastItemUsed) + ":" //$NON-NLS-1$
-                            + lastItemUsed.getProperty().getLabel();
-                }
-                if (item instanceof FileItem) { // hywang add for 6484
-                    lastFileItemUsed = (FileItem) item;
-                    if (lastFileItemUsed instanceof RulesItem) {
-                        displayName = "Rules:" //$NON-NLS-1$
-                                + lastFileItemUsed.getProperty().getLabel();
-                    }
-                }
-                if (item instanceof LinkRulesItem) {
-                    lastLinkItem = (LinkRulesItem) item;
-                    displayName = "Rules:" //$NON-NLS-1$
-                            + lastLinkItem.getProperty().getLabel();
-
-                }
-            } catch (PersistenceException e) {
-                ExceptionHandler.process(e);
-            }
-
-            param.setListItemsDisplayName(new String[] { displayName });
-            param.setListItemsValue(new String[] { (String) param.getValue() });
-        }
-
-    }
-
-    private void fastRepositoryUpdateSchema(IElementParameter param) {
-        if (param != null && param.getValue() != null) {
-            String queryIdAndName = (String) param.getValue();
-            String[] names = queryIdAndName.split(" - "); //$NON-NLS-1$
-            if (names.length < 2) {
-                return;
-            }
-            String linkedRepository = names[0];
-            String tableName = null;
-            if (names.length == 2) {
-                tableName = names[1];
-            } else if (names.length > 2) {
-                tableName = queryIdAndName.substring(linkedRepository.length() + 3);
-            }
-
-            if (lastItemUsed != null) {
-                if (!linkedRepository.equals(lastItemUsed.getProperty().getId())) {
-                    lastItemUsed = null;
-                }
-            }
-            if (lastItemUsed == null) {
-                IProxyRepositoryFactory factory = DesignerPlugin.getDefault().getProxyRepositoryFactory();
-                Item item;
-                try {
-                    IRepositoryViewObject object = factory.getLastVersion(linkedRepository);
-                    if (object == null) {
-                        return;
-                    }
-                    item = object.getProperty().getItem();
-                    Assert.isTrue(item instanceof ConnectionItem);
-                    lastItemUsed = (ConnectionItem) item;
-                } catch (PersistenceException e) {
-                    ExceptionHandler.process(e);
-                }
-            }
-
-            // EList<MetadataTable> tableList = lastItemUsed.getConnection().getTables();
-            // for (MetadataTable table )
-
-            String displayName = dynamicProperty.getRepositoryAliasName(lastItemUsed) + ":" //$NON-NLS-1$
-                    + lastItemUsed.getProperty().getLabel() + " - " + tableName; //$NON-NLS-1$
-            IElementParameter infoObjectTypeParam = param.getElement().getElementParameter("INFO_OBJECT_TYPE"); //$NON-NLS-1$
-            if (infoObjectTypeParam != null) {
-                String innerIOType = (String) infoObjectTypeParam.getValue();
-                if (innerIOType != null) {
-                    displayName = displayName + " (" + innerIOType + ")"; //$NON-NLS-1$ //$NON-NLS-2$
-                }
-            }
-            param.setListItemsDisplayName(new String[] { displayName });
-            param.setListItemsValue(new String[] { (String) param.getValue() });
-        }
-    }
-
-    private void fastRepositoryUpdateQuery(IElementParameter param) {
-        if (param != null && param.getValue() != null) {
-            String queryIdAndName = (String) param.getValue();
-            String[] names = queryIdAndName.split(" - "); //$NON-NLS-1$
-            if (names.length < 2) {
-                return;
-            }
-            String linkedRepository = names[0];
-            String queryName = null;
-            if (names.length == 2) {
-                queryName = names[1];
-            } else if (names.length > 2) {
-                queryName = queryIdAndName.substring(linkedRepository.length() + 3);
-            }
-            if (lastItemUsed != null) {
-                if (!linkedRepository.equals(lastItemUsed.getProperty().getId())) {
-                    lastItemUsed = null;
-                }
-            }
-            if (lastItemUsed == null) {
-                IProxyRepositoryFactory factory = DesignerPlugin.getDefault().getProxyRepositoryFactory();
-                Item item;
-                try {
-                    IRepositoryViewObject object = factory.getLastVersion(linkedRepository);
-                    if (object == null) {
-                        return;
-                    }
-                    item = object.getProperty().getItem();
-                    lastItemUsed = (ConnectionItem) item;
-                } catch (PersistenceException e) {
-                    ExceptionHandler.process(e);
-                }
-            }
-            Assert.isTrue(lastItemUsed instanceof DatabaseConnectionItem);
-            QueriesConnection queriesConnection = ((DatabaseConnection) lastItemUsed.getConnection()).getQueries();
-            EList<Query> queries = queriesConnection.getQuery();
-
-            String repositoryAliasName = dynamicProperty.getRepositoryAliasName(lastItemUsed);
-            for (Query currentQuery : queries) {
-                if (currentQuery.getLabel().equals(queryName)) {
-                    String displayName = repositoryAliasName + ":" //$NON-NLS-1$
-                            + lastItemUsed.getProperty().getLabel() + " - " + currentQuery.getLabel(); //$NON-NLS-1$
-
-                    param.setListItemsDisplayName(new String[] { displayName });
-                    param.setListItemsValue(new String[] { (String) param.getValue() });
-                    /* query cache should be deleted ,bug 16969 */
-                    // dynamicProperty.getRepositoryQueryStoreMap().clear();
-                    // dynamicProperty.getRepositoryQueryStoreMap().put((String) param.getValue(), currentQuery);
-                }
-            }
-        }
-    }
-
-    protected void fastInitializeRepositoryNames(IElementParameter param) {
-        lastItemUsed = null;
-
-        if (param.getValue() == null || param.getValue().equals("")) {
-            return;
-        }
-        if (param.getName().equals(EParameterName.REPOSITORY_PROPERTY_TYPE.getName())) {
-            fastRepositoryUpdateProperty(param);
-        } else if (param.getName().equals(EParameterName.REPOSITORY_SCHEMA_TYPE.getName())) {
-            fastRepositoryUpdateSchema(param);
-        } else if (param.getName().equals(EParameterName.REPOSITORY_QUERYSTORE_TYPE.getName())) {
-            fastRepositoryUpdateQuery(param);
-        }
-    }
-
     protected String getDisplayNameFromValue(IElementParameter param, String value) {
-        // to load informations from repository only if needed.
-
-        int index = param.getIndexOfItemFromList(value);
-        IElementParameter infoObjectParam = param.getElement().getElementParameter("INFO_OBJECT_TYPE"); //$NON-NLS-1$
-        if (infoObjectParam != null && infoObjectParam.getValue() != null) {
-            index = -1;
-        }
-        if (index == -1) {
-
-            fastInitializeRepositoryNames(param);
-            // if even after the initialize there is nothing, just return an empty string
-            if (param.getListItemsDisplayName().length == 0) {
-                return ""; //$NON-NLS-1$
-            }
-            index = param.getIndexOfItemFromList(value);
-            if (index == -1) {
-                return ""; //$NON-NLS-1$
-            }
-        }
-
-        return param.getListItemsDisplayName()[index];
+        return repositoryValueHander.getDisplayNameFromValue(param, value);
     }
 
     /*
@@ -585,4 +376,12 @@ public abstract class AbstractRepositoryController extends AbstractElementProper
 
     }
 
+    /**
+     * Getter for repositoryValueHander.
+     * 
+     * @return the repositoryValueHander
+     */
+    public ControllerRepositoryValueHander getRepositoryValueHander() {
+        return this.repositoryValueHander;
+    }
 }
